fix bug

+++ src/Components/DeleteInfo.jsx
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from 'react';
 import '../Styles/content.css';
 import { motion } from "framer-motion"
 import axios from 'axios';
-import { Link } from 'react-router-dom';
 
 const DeleteInfo = ({ tipe }) => {
     //? state for controlling the ui showed in the client-side.
@@ -118,26 +117,18 @@ const DeleteInfo = ({ tipe }) => {
                             <p className='isi text-sm text-justify px-4 py-4' dangerouslySetInnerHTML={{ __html: isi_2[idx] }}></p>
                             {item.gambar_2 ? <p className='font-semibold text-sm text-left px-4'><a className='underline text-blue-500' href={item.gambar_2}>{item.gambar_2}</a></p> : <></>}
                             <hr className='px-2 bg-blackmy-4 mx-auto w-48 h-1 bg-gray-100 rounded border-0 md:my-10 dark:bg-gray-700'></hr>
-                            {tipe === 'Artikel' ?
-                                <button
-                                    className='ml-[80%] my-4 mx-4 rounded-md bg-red-500 text-white p-4 text-sm font-semibold hover:bg-red-600'
-                                    onClick={() => {
-                                        const confirm = window.confirm('Apakah Anda Ingin Menghapus Data ini?')
-                                        if (confirm) {
+                            <button
+                                className='ml-[80%] my-4 mx-4 rounded-md bg-red-500 text-white p-4 text-sm font-semibold hover:bg-red-600'
+                                onClick={() => {
+                                    const confirm = window.confirm('Apakah Anda Ingin Menghapus Data ini?')
+                                    if (confirm) {
+                                        if (tipe === 'Artikel') {
                                             handleDeleteArtikel(item.id);
-                                        }
-                                    }}>Hapus Data</button>
-                                :
-                                <button
-                                    Link
-                                    className='ml-[80%] my-4 mx-4 rounded-md bg-red-500 text-white p-4 text-sm font-semibold hover:bg-red-600'
-                                    onClick={() => {
-                                        const confirm = window.confirm('Apakah Anda Ingin Menghapus Data ini?')
-                                        if (confirm) {
+                                        } else {
                                             handleDeleteBerita(item.id);
                                         }
-                                    }}><Link to={'/Beranfa'}>Hapus Data</Link></button>
-                            }
+                                    }
+                                }}>Hapus Data</button>
                         </div>
                     </motion.div>
                 ))
